TextExtraction/requirements.txt
pdf2image
requests
Pillow
pandas
groq
openpyxl
rapidfuzz
opencv-python
numpy
torch
clip-by-openai
fastapi
uvicorn
python-multipart